app/requirements.txt
numpy
Flask==0.12.2
Flask-Cors==3.0.3
Flask-RESTful==0.3.6
py-healthcheck
pdfminer.six==20170720
pdfplumber==0.5.8
PyPDF2==1.26.0
six==1.11.0
urllib3==1.22
Werkzeug==0.12.2
opencv-python
sortedcontainers
beautifulsoup4
requests
gunicorn
google-cloud-translate==1.3.1
pandas
python-docx
natsort